feat: move `get_all_messages/0` into `PsMessages` context
