+++ lib/screenplay/pa_messages.ex
@@ -3,9 +3,23 @@ defmodule Screenplay.PaMessages do
   Context module for functions dealing with `PaMessage`s
   """
 
+  # Use sparingly!
+  import Ecto.Query
+
   alias Screenplay.Alerts.Cache, as: AlertsCache
   alias Screenplay.PaMessages.PaMessage
   alias Screenplay.Repo
+
+  @doc """
+  Returns a list of ALL PA Messages ordered by their inserted_at timestamps
+  descending.
+  """
+  @spec get_all_messages() :: [PaMessage.t()]
+  def get_all_messages do
+    PaMessage
+    |> order_by(desc: :inserted_at)
+    |> Repo.all()
+  end
 
   @doc """
   Returns a list of the currently active PA Messages.
